requirements.txt
datetime
imageio
numpy
configparser
tqdm
astropy
matplotlib
tensorflow-gpu
photutils
seaborn
scipy